requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pillow==11.2.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.6.1
scikit-surprise==1.1.4
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
